perf: using long for price istead int

+++ app/src/main/java/com/catcode/note_app/ui/dialog/AddNoteDialog.kt
@@ -86,9 +86,9 @@ class AddNoteDialog(
           return@setOnClickListener
         }
 
-        val rawPrice = CurrencyFormatter.extractRawValue(
-          inputPrice.text.toString()
-        )
+        val rawPrice = CurrencyFormatter
+          .extractRawValue(inputPrice.text.toString())
+          .toLong()
 
         val note = NoteEntity(
           name = name,
